DO NOT MERGE - remove Client.remotes
The serverName -> *Remote mapping is a LRU, even if there won't be many
serverNames. Instead the SKI -> *Remote mapping is a simple map, from
which nothing is never removed. It will unavoidably grow in size the
more certs are observed.

One could obviously just use an LRU, but it got me thinking about why we
have that indirection anyway, when we could instead just stick the
serverName on the PrivateKey, and rely on the serverName -> *Remote LRU.

A lot of removed lines seem to suggest this simplifies things.

This is NOT a finished patch (haven't touched the tests), but it's
easier to talk over code.