Don't ask for user input by default
The default behavior of the keyless server when the server.pem and
server-key.pem files are not present is to ask the user for input to
generate them. But if the server was launched as a daemon, it will
just be stuck waiting for input, while appearing to be running.

With this change, the user will not be prompted for input unless the
-config-only or -manual-activation flags are provided.