feat: add opentracing + jaeger
This adds support for opentracing with Jaeger. It is disabled by default, and can be turned on with a configurable sample rate. The bulk of the changes here are adding `context.Context` as the first parameter to the internal functions to allow for span context propogation.

This does change the protocol, adding a new optional field to allow for span propogation between keyless instances. https://github.com/cloudflare/gokeyless/pull/276 allows this.